+++ services/ui-src/src/measures/2023/shared/util/validationsMock.tsx
@@ -6,7 +6,7 @@ import * as validateAtLeastOneDeviationFieldFilled from "measures/2023/shared/gl
 import * as validateAtLeastOneRateComplete from "shared/globalValidations/validateAtLeastOneRateComplete";
 import * as validateBothDatesInRange from "shared/globalValidations/validateBothDatesInRange";
 import * as validateDualPopInformation from "measures/2023/shared/globalValidations/validateDualPopInformation";
-import * as validateEqualCategoryDenominators from "measures/2023/shared/globalValidations/validateEqualCategoryDenominators";
+import * as validateEqualCategoryDenominators from "shared/globalValidations/validateEqualCategoryDenominators";
 import * as validateEqualQualifierDenominators from "shared/globalValidations/validateEqualQualifierDenominators";
 import * as validateFfsRadioButtonCompletion from "measures/2023/shared/globalValidations/validateFfsRadioButtonCompletion";
 import * as validateRateNotZero from "measures/2023/shared/globalValidations/validateRateNotZero";
