+++ services/ui-src/src/shared/globalValidations/validateEqualCategoryDenominators/index.test.ts
